requirements.txt
faster-whisper==0.10.0
streamlit==1.29.0